refactor(gql): move Filmstaden related queries
to its own file.
