+++ internal/graph/movies.resolvers.go
@@ -125,56 +125,3 @@ func (r *queryResolver) ArchivedMovies(ctx context.Context) ([]*model.Movie, err
 	}
 	return movies, nil
 }
-
-func (r *queryResolver) FilmstadenShowings(ctx context.Context, movieID uuid.UUID, city *string, afterDate *time.Time) ([]*model.FilmstadenShowing, error) {
-	var cityAlias string
-	var afterTime time.Time
-	if city == nil {
-		cityAlias = "GB"
-	} else {
-		cityAlias = *city
-	}
-	if afterDate == nil {
-		afterTime = time.Unix(0, 0)
-	} else {
-		afterTime = *afterDate
-	}
-
-	logger := logging.FromContext(ctx).
-		WithValues(
-			"movieID", movieID,
-			"cityAlias", cityAlias,
-			"afterDate", afterDate,
-		)
-
-	fsMovieID, err := database.FromContext(ctx).LookupFilmstadenID(ctx, movieID)
-	if err != nil {
-		logger.Error(err, "query.LookupFilmstadenID failed")
-		return nil, fmt.Errorf("failed to lookup FS movie ID")
-	}
-
-	shows, err := r.filmstaden.ShowsForMovie(ctx, 1, cityAlias, fsMovieID, afterTime)
-	if err != nil {
-		logger.Error(err, "filmstaden.ShowsForMovie failed")
-		return nil, fmt.Errorf("failed to fetch shows from Filmstaden")
-	}
-
-	fsShows := make([]*model.FilmstadenShowing, len(shows.Items))
-	for i, show := range shows.Items {
-		fsShows[i] = &model.FilmstadenShowing{
-			ID: show.RemoteEntityID,
-			Cinema: &model.FilmstadenCinema{
-				FilmstadenID: show.CinemaID,
-				Name:         show.CinemaTitle,
-			},
-			Screen: &model.FilmstadenScreen{
-				FilmstadenID: show.ScreenID,
-				Name:         show.ScreenTitle,
-			},
-			TimeUtc: show.UTC,
-			Tags:    show.ScreenAttributes,
-		}
-	}
-
-	return fsShows, nil
-}
